requirements.txt
streamlit
pandas
pdfplumber
python-docx
sentence-transformers
torch
transformers
